Code refactor

+++ common/Serialization.cpp
@@ -530,19 +530,19 @@ std::pair<size_t, uint64_t> Deserialization::getVarint(off_t offset) const
 std::string Deserialization::getString(off_t offset, size_t size) const
 {
     WCTInnerAssert(size > 0);
-    WCTInnerAssert(isEnough(offset + size));
+    WCTInnerAssert(isEnough((size_t) offset + size));
     return std::string(reinterpret_cast<const char *>(base() + offset), size);
 }
 
 const UnsafeData Deserialization::getData(off_t offset, size_t size) const
 {
-    WCTInnerAssert(isEnough(offset + size));
+    WCTInnerAssert(isEnough((size_t) offset + size));
     return UnsafeData::immutable(base() + offset, size);
 }
 
 int64_t Deserialization::get8BytesInt(off_t offset) const
 {
-    WCTInnerAssert(isEnough(offset + 8));
+    WCTInnerAssert(isEnough((size_t) offset + 8));
     uint64_t x = get4BytesUInt(offset);
     uint32_t y = get4BytesUInt(offset + 4);
     x = (x << 32) + y;
@@ -551,40 +551,40 @@ int64_t Deserialization::get8BytesInt(off_t offset) const
 
 int64_t Deserialization::get6BytesInt(off_t offset) const
 {
-    WCTInnerAssert(isEnough(offset + 6));
+    WCTInnerAssert(isEnough((size_t) offset + 6));
     return get4BytesInt(offset + 2) + (((int64_t) 1) << 32) * get2BytesInt(offset);
 }
 
 int32_t Deserialization::get4BytesInt(off_t offset) const
 {
-    WCTInnerAssert(isEnough(offset + 4));
+    WCTInnerAssert(isEnough((size_t) offset + 4));
     const unsigned char *p = base() + offset;
     return (16777216 * (int8_t)(p[0]) | (p[1] << 16) | (p[2] << 8) | p[3]);
 }
 
 int32_t Deserialization::get3BytesInt(off_t offset) const
 {
-    WCTInnerAssert(isEnough(offset + 3));
+    WCTInnerAssert(isEnough((size_t) offset + 3));
     const unsigned char *p = base() + offset;
     return (65536 * (int8_t)(p[0]) | (p[1] << 8) | p[2]);
 }
 
 int32_t Deserialization::get2BytesInt(off_t offset) const
 {
-    WCTInnerAssert(isEnough(offset + 2));
+    WCTInnerAssert(isEnough((size_t) offset + 2));
     const unsigned char *p = base() + offset;
     return (256 * (int8_t)(p[0]) | p[1]);
 }
 
 int32_t Deserialization::get1ByteInt(off_t offset) const
 {
-    WCTInnerAssert(isEnough(offset + 1));
+    WCTInnerAssert(isEnough((size_t) offset + 1));
     return (int8_t)(base() + offset)[0];
 }
 
 double Deserialization::get8BytesDouble(off_t offset) const
 {
-    WCTInnerAssert(isEnough(offset + 8));
+    WCTInnerAssert(isEnough((size_t) offset + 8));
     uint64_t x = get4BytesUInt(offset);
     uint32_t y = get4BytesUInt(offset + 4);
     x = (x << 32) + y;
@@ -595,7 +595,7 @@ double Deserialization::get8BytesDouble(off_t offset) const
 
 uint32_t Deserialization::get4BytesUInt(off_t offset) const
 {
-    WCTInnerAssert(isEnough(offset + 4));
+    WCTInnerAssert(isEnough((size_t) offset + 4));
     const unsigned char *p = base() + offset;
     return (((uint32_t) p[0] << 24) | (p[1] << 16) | (p[2] << 8) | p[3]);
 }
